+++ app/Livewire/VinList.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Livewire;
+
+use Livewire\Component;
+
+class VinList extends Component
+{
+    public function render()
+    {
+        return view('livewire.vin-list');
+    }
+}
