+++ composer.json
@@ -12,6 +12,7 @@
         "livewire/livewire": "^3.2"
     },
     "require-dev": {
+        "barryvdh/laravel-debugbar": "^3.9",
         "fakerphp/faker": "^1.9.1",
         "laravel/pint": "^1.0",
         "laravel/sail": "^1.26",
