+++ config/cache.php
@@ -33,6 +33,10 @@ return [
 
     'stores' => [
 
+        'none' => [
+            'driver' => 'null',
+        ],
+
         'apc' => [
             'driver' => 'apc',
         ],
