+++ config/view.php
@@ -28,10 +28,10 @@ return [
     |
     */
 
-/*    'compiled' => env(
+    /*    'compiled' => env(
         'VIEW_COMPILED_PATH',
         realpath(storage_path('framework/views'))
     ),*/
-    'compiled' => '/tmp'
+    'compiled' => '/tmp',
 
 ];
