+++ resources/views/livewire/vin-list.blade.php
@@ -1,0 +1,4 @@
+<div>
+    {{-- If you look to others for fulfillment, you will never truly be fulfilled. --}}
+    <h1>
+</div>
