+++ resources/views/livewire/vin-show.blade.php
@@ -1,0 +1,3 @@
+<div>
+    {{-- Be like water. --}}
+</div>
